+++ src/config/chis-ug/index.ts
@@ -1,10 +1,8 @@
 import { PartnerConfig } from '..';
 import config from './config.json';
-import mutate from './gross';
 
 const partnerConfig: PartnerConfig = {
   config,
-  mutate,
 };
 
 export default partnerConfig;
